again add some stronger assumptions to end-to-end lemma
the non-passive and passive global states must coincide on the constants, which means the constants have to be constrained at the beginning. This assumption was removed in an earlier commit but now has been added in a new way. Instead of forcing the variable relation to map every constant, we directly encode the assumption in the construction of the initial set.