fixed bug in fun_interp_wf_2 (artifact is not affected)
The LHS of implication is always guaranteed when we unfold the definition and we currently only use "type_of_val A v = instantiate ts ret_ty" (i.e., our proofs require that if a function reduces, then the output type is correct. However they don't require that if the function reduces then the inputs must have been well-typed).