Fix typing of reals, make typing tactic work for reals, and minor adjustments
Previously, the typing rules forced relational operators to be applicable only for integers. Now, the typing rules also permit reals (but never a mix of reals and integers, since such combinations are not well-typed in Boogie). Moreover, the typing rules are now expressed in a way that do not treat real division differently from integer division. As a result of this change, the typing tactic for real division should work (while previously it did not).